+++ Editor/Editors/TextureEditor/ChannelSelectEffect.cs
@@ -1,4 +1,5 @@
 ﻿using Editor.Utilities;
+using System.Threading.Channels;
 using System.Windows;
 using System.Windows.Media;
 using System.Windows.Media.Effects;